rencana pembangunan

+++ app/Http/Controllers/KomponenPembangunanController.php
@@ -21,10 +21,11 @@ class KomponenPembangunanController extends Controller
         $data = $request->all();
 
         $simpan =   KomponenPembangunan::create($data);
-
+        $latest_data = KomponenPembangunan::find($simpan->id);
         if ($simpan) {
             return response()->json([
-                'status' => 'success'
+                'status' => 'success',
+                'latest_data' => $latest_data
             ]);
         }
     }
@@ -34,18 +35,21 @@ class KomponenPembangunanController extends Controller
         $id = $request->id;
         $komponen = KomponenPembangunan::find($id);
 
-        $html = '<td>
-                    <span class="d-none">1000</span><i class="fas fa-window-close text-danger m-1" role="button" title="Batal" onclick="clearForm()"></i>
-                </td>
-                <td>
-                    <input type="hidden" id="id" name="id" value="'.$komponen->id.'">
-                    <input type="text" class="form-control form-control-sm" id="komponen" name="komponen" placeholder="Komponen pekerjaan" value="'.$komponen->komponen.'">
-                </td>
-                <td>
-                    <div class="d-flex justify-content-center">
-                        <i class="fas fa-save text-center" role="button" title="Save" onclick="updateKomponenPembangunan('.$komponen->id.')"></i>
-                    </div>
-                </td>';
+        $html = '
+                <tr class="form-input-data">
+                    <td>
+                        <span class="d-none">1000</span><i class="fas fa-window-close text-danger m-1" role="button" title="Batal" onclick="clearForm()"></i>
+                    </td>
+                    <td>
+                        <input type="hidden" id="id" name="id" value="'.$komponen->id.'">
+                        <input type="text" class="form-control form-control-sm" id="komponen" name="komponen" placeholder="Komponen pekerjaan" value="'.$komponen->komponen.'">
+                    </td>
+                    <td>
+                        <div class="d-flex justify-content-center">
+                            <i class="fas fa-save text-center text-success" role="button" title="Save" onclick="updateKomponenPembangunan('.$komponen->id.', event)"></i>
+                        </div>
+                    </td>
+                </tr>';
 
         if ($komponen->count() > 0) {
             return response()->json([
@@ -78,8 +82,15 @@ class KomponenPembangunanController extends Controller
         $id = $request->id;
 
         $komponen = KomponenPembangunan::find($id);
+        if ($komponen->parent == 0) {
+            $hapus = KomponenPembangunan::where('parent', $id)->delete();
+            if ($hapus) {
+                $hapus =    $komponen->delete();
+            }
+        }else {
+            $hapus =    $komponen->delete();
+        }
 
-        $hapus =    $komponen->delete();
 
         if ($hapus) {
             return response()->json([
